+++ CHANGELOG.md
@@ -773,3 +773,4 @@
 * Change `Zipper` of `Rose` tree definition
 * Define precedence and fixity for `@>>>` operator
 * Define `unfasten` method in Zippable` typeclass
+* Define `probably` method of `State` effect